requirements.txt
httpx>=0.27,<1.0
beautifulsoup4>=4.12
lxml>=5.0
python-dotenv>=1.0
tenacity>=8.2
click>=8.1
tqdm>=4.66
